src/requirements.txt
argparse~=1.4.0
ffmpeg-python~=0.2.0
Pyrogram~=2.0.0
distro~=1.9.0
curl_cffi~=0.12.0
requests~=2.32.0
ruff~=0.12.0
pre-commit~=4.3.0
